Refactoring: python_page.py -> python_main_page.py and it separated on two pages. Also README.md and requirements.txt changed.
