+++ requirements.txt
@@ -1,20 +1,34 @@
-allure-pytest==2.8.11
-allure-python-commons==2.8.11
+allure-pytest==2.5.4
+allure-python-commons==2.5.4
+altgraph==0.10.2
 appdirs==1.4.3
 attrs==19.3.0
-distlib==0.3.0
+cffi==1.14.0
+chardet==3.0.4
+contextlib2==0.6.0.post1
+decorator==4.4.2
 dnspython==1.16.0
 filelock==3.0.12
-more-itertools==8.2.0
+idna==2.9
+inflection==0.3.1
+javaobj-py3==0.4.0.1
 mysql-connector-python==8.0.19
-packaging==20.1
 pluggy==0.13.1
-protobuf==3.11.3
 py==1.8.1
-pycodestyle==2.5.0
-pyparsing==2.4.6
-pytest==5.3.5
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.20
+pycryptodome==3.9.7
+python-consul==1.1.0
+python-dateutil==2.8.1
 PyYAML==5.3
+requests==2.23.0
+rsa==4.0
+schematics==2.1.0
+selenium==3.141.0
 six==1.14.0
-virtualenv==20.0.7
-wcwidth==0.1.8
+text-unidecode==1.3
+toml==0.10.0
+twofish==0.3.0
+urllib3==1.25.8
+virtualenv==20.0.0
