+++ ui_testing/tests/test_python_docs.py
@@ -1,4 +1,5 @@
-from ui_testing.pages.python_page import PythonMainPage
+from ui_testing.pages.docs_page import DocsPage
+from ui_testing.pages.python_main_page import PythonMainPage
 import ui_testing.service as service
 
 
@@ -11,17 +12,19 @@ def test_check_python_metaclasses_history(browser):
 
     python_page.click_docs_link()
 
-    title = python_page.get_page_title()
+    docs_page = DocsPage(browser)
+
+    title = docs_page.get_page_title()
     assert title == service.get_expected_title("documentation_page_title")
 
-    python_page.search_doc(service.get_required_search_text())
+    docs_page.search_doc(service.get_required_search_text())
 
-    python_page.get_doc_for_metaclasses()
+    docs_page.get_doc_for_metaclasses()
 
-    title = python_page.get_page_title()
+    title = docs_page.get_page_title()
     assert title == service.get_expected_title("doc_for_metaclasses_page_title")
 
-    actual = python_page.get_post_history_content()
+    actual = docs_page.get_post_history_content()
     expected = service.get_expected_post_history()
 
     assert expected == actual
